requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-core==2.14.0
deep-translator==1.11.4
colorama==0.4.6
openai==1.3.0
aiohttp==3.9.1
certifi==2023.11.17
python-multipart==0.0.6
asyncio==3.4.3
ssl==1.16.0